requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
GitPython==3.1.43
PyYAML==6.0.2
python-dotenv==1.0.1
litellm==1.52.12

